Some progress on measuring primes
Ran into a challenge where the same method would have an order of
magnitude difference in performance, depending on how it was measured.
`ch1_primes2.rkt` is a start to be more systematic about measurement
across different scenarios, but it will require more syntactical tools
than we've been taught about Lisps so far in the book.